requirements.txt
keras
matplotlib
numpy
pillow
pyqt5
seaborn
scikit-learn
tensorflow
tqdm